requirements.txt
PyPDF2>=3.0.0
tqdm>=4.0.0

# Embeddings and search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
numpy>=1.24.0
streamlit>=1.22.0
Flask>=2.2.5
